Refactoring aframe components
