+++ src/aframe/click-open-modal.js
@@ -1,0 +1,7 @@
+AFRAME.registerComponent('click-open-modal', {
+  init() {
+    this.el.addEventListener('click', () => {
+      document.querySelector('#modal').style.display = 'flex';
+    });
+  },
+});
